requirements.txt
fastapi
uvicorn
pytesseract
pdf2image
Pillow
